refactor(config): Separate dataclasses into schemas.py
